+++ bot.py
@@ -1,0 +1,15 @@
+import discord
+from casinogames import *
+
+"""
+DISCORD BOT STARTUP SCREEN
+"""
+@client.event
+async def on_ready():
+    print('Logged in as')
+    print(client.user.name)
+    print(client.user.id)
+    print ('Bot working properly')
+    print('------')
+
+client.run('MTg3NjYwNDA4MDIxMDU3NTM3.CuKrag.9X9myjLSYD2J9IX6ANWal4ZqPNM')
